modifications : created interface for AttendanceService
