+++ src/main/java/com/esad/emasys/services/AttendanceService.java
@@ -1,83 +1,13 @@
 package com.esad.emasys.services;
 
-import com.esad.emasys.impl.EmployeeServiceImpl;
-import com.esad.emasys.model.Attendance;
 import com.esad.emasys.model.AttendanceStatusResponse;
-import com.esad.emasys.model.Employee;
-import com.esad.emasys.repository.AttendanceRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-import java.time.Duration;
 import java.time.LocalDateTime;
-import java.util.Optional;
 
-@Service
-public class AttendanceService {
+public interface AttendanceService {
+    LocalDateTime checkIn(Integer empId);
 
-    @Autowired
-    private AttendanceRepository attendanceRepository;
+    LocalDateTime checkOut(Integer empId);
 
-    @Autowired
-    private EmployeeServiceImpl empService;
-
-    public LocalDateTime checkIn(Integer empId) {
-        Employee emp = empService.getEmployee(empId);
-
-        // Check if the user has already checked in and hasn't checked out yet
-        if (attendanceRepository.findByEmployeeAndCheckOutTimeIsNull(emp).isPresent()) {
-            throw new IllegalStateException("User has already checked in but not checked out yet");
-        }
-
-        Attendance attendance = new Attendance();
-        attendance.setEmployee(emp);
-        attendance.setCheckInTime(LocalDateTime.now());
-        attendance.setAttendanceDate(LocalDateTime.now());
-        attendanceRepository.save(attendance);
-
-        return attendance.getCheckInTime();
-    }
-
-    public LocalDateTime checkOut(Integer empId) {
-        Employee emp = empService.getEmployee(empId);
-
-        Attendance attendance = attendanceRepository.findByEmployeeAndCheckOutTimeIsNull(emp)
-                .orElseThrow(() -> new IllegalStateException("User has not checked in yet"));
-
-        LocalDateTime checkOutTime = LocalDateTime.now();
-        attendance.setCheckOutTime(checkOutTime);
-        attendance.setStatus(Attendance.Status.CHECK_OUT);
-
-        // Calculate the total worked hours
-        LocalDateTime checkInTime = attendance.getCheckInTime();
-        Duration duration = Duration.between(checkInTime, checkOutTime);
-        float totalHours = duration.toMinutes() / 60.0f;  // Convert minutes to hours
-
-        // Set the total hours worked
-        attendance.setTotalHours(totalHours);
-        attendanceRepository.save(attendance);
-
-        return attendance.getCheckOutTime();
-    }
-
-    public AttendanceStatusResponse getStatus(int employeeId) {
-        // Step 1: Find the latest CHECK_IN with totalHours = 0
-        Optional<Attendance> checkInOpt = attendanceRepository
-                .findFirstByEmployeeIdAndStatusOrderByCheckInTimeDesc(employeeId, Attendance.Status.CHECK_IN);
-
-        Optional<Attendance> checkOutOpt = attendanceRepository
-                .findFirstByEmployeeIdAndStatusOrderByCheckOutTimeDesc(employeeId, Attendance.Status.CHECK_OUT);
-
-        if (checkInOpt.isPresent()) {
-            // Step 2: If CHECK_IN found with totalHours = 0, return it
-            Attendance attendance = checkInOpt.get();
-            return new AttendanceStatusResponse(attendance.getStatus(), attendance.getCheckInTime(), null, attendance.getAttendanceDate(), attendance.getTotalHours());
-        } else if (checkOutOpt.isPresent()) {
-            // Step 3: If no CHECK_IN found, find the latest CHECK_OUT
-            Attendance attendance = checkOutOpt.get();
-            return new AttendanceStatusResponse(attendance.getStatus(), null, attendance.getCheckOutTime(), attendance.getAttendanceDate(), attendance.getTotalHours());
-        } else {
-            throw null;
-        }
-    }
+    AttendanceStatusResponse getStatus(Integer empId);
 }
